Move parse sdr function to ipmi module
The sdr info are retrieved using ipmitool, the function should stay in the
ipmi module.
Also refactor it to not pass the hardware info list around.
