+++ hardware/tests/results/detect_utils_results.py
@@ -11,96 +11,6 @@
 # under the License.
 
 
-IPMI_SDR_RESULT = [('ipmi', 'UID Light', 'value', '0x00'),
-                   ('ipmi', 'Sys. Health LED', 'value', '0x00'),
-                   ('ipmi', 'Power Supply 1', 'value', '90'),
-                   ('ipmi', 'Power Supply 1', 'unit', 'Watts'),
-                   ('ipmi', 'Power Supply 2', 'value', '65'),
-                   ('ipmi', 'Power Supply 2', 'unit', 'Watts'),
-                   ('ipmi', 'Power Supplies', 'value', '0x00'),
-                   ('ipmi', 'Fan 1', 'value', '33.32'),
-                   ('ipmi', 'Fan 1', 'unit', 'percent'),
-                   ('ipmi', 'Fan 2', 'value', '39.20'),
-                   ('ipmi', 'Fan 2', 'unit', 'percent'),
-                   ('ipmi', 'Fan 3', 'value', '39.20'),
-                   ('ipmi', 'Fan 3', 'unit', 'percent'),
-                   ('ipmi', 'Fan 4', 'value', '29.40'),
-                   ('ipmi', 'Fan 4', 'unit', 'percent'),
-                   ('ipmi', 'Fan 5', 'value', '24.70'),
-                   ('ipmi', 'Fan 5', 'unit', 'percent'),
-                   ('ipmi', 'Fan 6', 'value', '13.72'),
-                   ('ipmi', 'Fan 6', 'unit', 'percent'),
-                   ('ipmi', 'Fans', 'value', '0x00'),
-                   ('ipmi', 'Temp 1', 'value', '20'),
-                   ('ipmi', 'Temp 1', 'unit', 'degrees C'),
-                   ('ipmi', 'Temp 2', 'value', '40'),
-                   ('ipmi', 'Temp 2', 'unit', 'degrees C'),
-                   ('ipmi', 'Temp 3', 'value', '40'),
-                   ('ipmi', 'Temp 3', 'unit', 'degrees C'),
-                   ('ipmi', 'Temp 4', 'value', '28'),
-                   ('ipmi', 'Temp 4', 'unit', 'degrees C'),
-                   ('ipmi', 'Temp 5', 'value', '28'),
-                   ('ipmi', 'Temp 5', 'unit', 'degrees C'),
-                   ('ipmi', 'Temp 6', 'value', '34'),
-                   ('ipmi', 'Temp 6', 'unit', 'degrees C'),
-                   ('ipmi', 'Temp 7', 'value', '33'),
-                   ('ipmi', 'Temp 7', 'unit', 'degrees C'),
-                   ('ipmi', 'Temp 8', 'value', '39'),
-                   ('ipmi', 'Temp 8', 'unit', 'degrees C'),
-                   ('ipmi', 'Temp 9', 'value', '33'),
-                   ('ipmi', 'Temp 9', 'unit', 'degrees C'),
-                   ('ipmi', 'Temp 10', 'value', '39'),
-                   ('ipmi', 'Temp 10', 'unit', 'degrees C'),
-                   ('ipmi', 'Temp 11', 'value', '29'),
-                   ('ipmi', 'Temp 11', 'unit', 'degrees C'),
-                   ('ipmi', 'Temp 12', 'value', '40'),
-                   ('ipmi', 'Temp 12', 'unit', 'degrees C'),
-                   ('ipmi', 'Temp 13', 'value', '28'),
-                   ('ipmi', 'Temp 13', 'unit', 'degrees C'),
-                   ('ipmi', 'Temp 14', 'value', '31'),
-                   ('ipmi', 'Temp 14', 'unit', 'degrees C'),
-                   ('ipmi', 'Temp 15', 'value', '29'),
-                   ('ipmi', 'Temp 15', 'unit', 'degrees C'),
-                   ('ipmi', 'Temp 16', 'value', '25'),
-                   ('ipmi', 'Temp 16', 'unit', 'degrees C'),
-                   ('ipmi', 'Temp 17', 'value', '27'),
-                   ('ipmi', 'Temp 17', 'unit', 'degrees C'),
-                   ('ipmi', 'Temp 18', 'value', 'disabled'),
-                   ('ipmi', 'Temp 19', 'value', '22'),
-                   ('ipmi', 'Temp 19', 'unit', 'degrees C'),
-                   ('ipmi', 'Temp 20', 'value', '28'),
-                   ('ipmi', 'Temp 20', 'unit', 'degrees C'),
-                   ('ipmi', 'Temp 21', 'value', '28'),
-                   ('ipmi', 'Temp 21', 'unit', 'degrees C'),
-                   ('ipmi', 'Temp 22', 'value', '28'),
-                   ('ipmi', 'Temp 22', 'unit', 'degrees C'),
-                   ('ipmi', 'Temp 23', 'value', '33'),
-                   ('ipmi', 'Temp 23', 'unit', 'degrees C'),
-                   ('ipmi', 'Temp 24', 'value', '30'),
-                   ('ipmi', 'Temp 24', 'unit', 'degrees C'),
-                   ('ipmi', 'Temp 25', 'value', '30'),
-                   ('ipmi', 'Temp 25', 'unit', 'degrees C'),
-                   ('ipmi', 'Temp 26', 'value', '31'),
-                   ('ipmi', 'Temp 26', 'unit', 'degrees C'),
-                   ('ipmi', 'Temp 27', 'value', 'disabled'),
-                   ('ipmi', 'Temp 28', 'value', '26'),
-                   ('ipmi', 'Temp 28', 'unit', 'degrees C'),
-                   ('ipmi', 'Temp 29', 'value', '35'),
-                   ('ipmi', 'Temp 29', 'unit', 'degrees C'),
-                   ('ipmi', 'Temp 30', 'value', '58'),
-                   ('ipmi', 'Temp 30', 'unit', 'degrees C'),
-                   ('ipmi', 'Memory', 'value', '0x00'),
-                   ('ipmi', 'Power Meter', 'value', '170'),
-                   ('ipmi', 'Power Meter', 'unit', 'Watts'),
-                   ('ipmi', 'Cntlr 1 Bay 1', 'value', '0x01'),
-                   ('ipmi', 'Cntlr 1 Bay 2', 'value', '0x01'),
-                   ('ipmi', 'Cntlr 1 Bay 3', 'value', '0x00'),
-                   ('ipmi', 'Cntlr 1 Bay 4', 'value', '0x01'),
-                   ('ipmi', 'Cntlr 2 Bay 5', 'value', '0x00'),
-                   ('ipmi', 'Cntlr 2 Bay 6', 'value', '0x00'),
-                   ('ipmi', 'Cntlr 2 Bay 7', 'value', '0x01'),
-                   ('ipmi', 'Cntlr 2 Bay 8', 'value', '0x01')]
-
 GET_CPUS_RESULT = [('cpu', 'physical', 'number', 2),
                    ('cpu', 'physical_0', 'vendor', 'AuthenticAMD'),
                    ('cpu', 'physical_0', 'product',
